Update cquthesis.dtx
I read the article "How to Package Your LaTeX Package - of /tex-archive - CTAN" and found that checksum is simply set to be the number of backslash symbol in the .dtx file. Thus I think it should be changed from 1891 to 1893. 
The package can be successfully built on my machine (TeXLive 2020 on Ubuntu).